refactor(shared): реорганизация shared/, перенос модулей и фиксы импортов
- Перенесены:
  - lib/hooks/* -> hooks/
  - lib/api/* -> api/
  - lib/redux/* и model/app/* -> store/
  - lib/signUpSchema/* -> schemas/
  - lib/random/getRandomNumber.ts -> utils/
  - utils/localStorage/localStorage.ts -> utils/localStorage.ts
- Удалены пустые директории: lib/, model/
- Обновлены все импорты на новые пути
- Проект успешно собирается (next build)

BREAKING CHANGE: обновлены пути импортов внутри shared/*
