+++ src/features/auth/ui/Register/RegisterForm/RegisterForm.tsx
@@ -9,7 +9,7 @@ import { TextField } from "@/shared/ui/text-field/text-field"
 import { Button } from "@/shared/ui"
 import { Controller, useForm } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
-import { SignUpInputs, signUpSchema } from "@/shared/lib/signUpSchema/signUpSchema"
+import { SignUpInputs, signUpSchema } from "@/shared/schemas/signUpSchema"
 import { UserType } from "@/features/auth/api/authApi"
 import { useState } from "react"
 import Checkbox from "@/shared/ui/checkbox/checkbox"
