+++ src/features/auth/ui/SignIn/SignInForm.tsx
@@ -7,7 +7,7 @@ import githubIcon from "@/shared/assets/img/github-icon.png"
 // import { Button } from "@radix-ui/themes"
 import * as React from "react"
 import { useForm } from "react-hook-form"
-import { LoginInputs, loginSchema } from "@/shared/lib/signUpSchema/loginSchema"
+import { LoginInputs, loginSchema } from "@/shared/schemas/loginSchema"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { useState } from "react"
 import { TextField } from "@/shared/ui/text-field/text-field"
