+++ src/features/posts/ui/CroppingButton/CroppingButton.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, useRef, useState } from "react"
 import s from "./CroppingButton.module.css"
-import { useOutsideClick } from "@/shared/lib/hooks/useOutsideClick"
+import { useOutsideClick } from "@/shared/hooks/useOutsideClick"
 
 type Props = {
   hidden: ReactNode
