+++ src/features/posts/ui/PostModal/PostModal.module.scss
