+++ src/features/posts/ui/PostModal/index.ts
@@ -1,1 +1,1 @@
-export * from "./ui/PostModal"
+export * from "./PostModal"
