+++ src/features/posts/ui/StepsCreatePost/PostSettingModal.tsx
@@ -8,11 +8,11 @@ import { AddFotoStep } from "./AddFotoStep/AddFotoStep"
 import { CanvasImage, CroppingStep } from "./CroppingStep/CroppingStep"
 import { FilterStep } from "./FiltersStep/FiltersStep"
 import { PublicationStep } from "./PublicationStep/PublicationStep"
-import { useAppDispatch } from "@/shared/lib/redux/useAppDispatch"
+import { useAppDispatch } from "@/shared/store/useAppDispatch"
 import { addImageAC, addPostAC, changeImageAC, clearImagesAC, selectImages } from "@/features/posts/model/postsSlice"
-import { useAppSelector } from "@/shared/lib/redux/useAppSelector"
+import { useAppSelector } from "@/shared/store/useAppSelector"
 import { ModalWindow } from "@/features/auth/ui/Register/ModalWindow/ModalWindow"
-import { useOutsideClick } from "@/shared/lib/hooks/useOutsideClick"
+import { useOutsideClick } from "@/shared/hooks/useOutsideClick"
 import s from "./PostSettingModal.module.scss"
 import { useCreatePostMutation, useUploadImagesMutation } from "@/features/posts/api/postsApi"
