+++ src/widgets/auth/ui/SignUpContent/SignUpContent.tsx
@@ -5,7 +5,7 @@ import { Suspense } from "react"
 import { ModalWindow } from "@/features/auth/ui/Register/ModalWindow/ModalWindow"
 import { RegisterForm } from "@/features/auth/ui/Register/RegisterForm/RegisterForm"
 import { useConfirmationMutation, useRegistrationMutation, UserType } from "@/features/auth/api/authApi"
-import { useAppDispatch } from "@/shared/lib/redux/useAppDispatch"
+import { useAppDispatch } from "@/shared/store/useAppDispatch"
 import { useSearchParams } from "next/navigation"
 import { Confirmation } from "@/widgets/auth/ui/Confirmation/Confirmation"
 import Button from "@/shared/ui/button/button"
